Refactored saving/loading into a separate module
